+++ alembic/versions/e9c1224f6bb_add_fallback_flag_to_api_key.py
@@ -15,7 +15,9 @@ import sqlalchemy as sa
 
 
 def upgrade():
-    stmt = "ALTER TABLE api_key ADD COLUMN allow_fallback tinyint(1) AFTER description"
+    stmt = 'ALTER TABLE api_key ADD COLUMN allow_fallback tinyint(1) AFTER log'
+    op.execute(sa.text(stmt))
+    stmt = 'UPDATE api_key SET allow_fallback = 0 WHERE allow_fallback IS NULL'
     op.execute(sa.text(stmt))