+++ ichnaea/locate/stats.py
@@ -29,3 +29,7 @@ class StatsLogger(object):
     def stat_time(self, stat, count):
         self.stats_client.timing('{api}.{stat}'.format(
             api=self.api_name, stat=stat), count)
+
+    def stat_timer(self, stat):
+        return self.stats_client.timer('{api}.{stat}'.format(
+            api=self.api_name, stat=stat))